feat: add trigger edge cases
- Trigger are now combined in an OR statement
- Trigger actions are attached to their matching trigger
- Actions below triggers can have requirements
