+++ art-flow-parser/src/test/java/net/silthus/art/FlowParserTest.java
@@ -20,24 +20,19 @@ import com.google.inject.Provider;
 import lombok.SneakyThrows;
 import net.silthus.art.api.ArtContext;
 import net.silthus.art.api.ArtManager;
-import net.silthus.art.api.actions.ActionConfig;
-import net.silthus.art.api.actions.ActionContext;
 import net.silthus.art.api.config.ArtConfig;
-import net.silthus.art.api.config.ArtObjectConfig;
 import net.silthus.art.api.parser.ArtParseException;
-import net.silthus.art.api.requirements.RequirementConfig;
-import net.silthus.art.api.requirements.RequirementContext;
-import net.silthus.art.api.trigger.TriggerConfig;
-import net.silthus.art.api.trigger.TriggerContext;
 import net.silthus.art.parser.flow.parser.ArtTypeParser;
-import org.assertj.core.groups.Tuple;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.mockito.stubbing.Answer;
 
-import java.util.*;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.Set;
 import java.util.regex.Pattern;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -45,7 +40,7 @@ import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.mockito.Mockito.*;
 
 @DisplayName("FlowParser")
-@SuppressWarnings("unchecked")
+@SuppressWarnings({"unchecked", "rawtypes"})
 class FlowParserTest {
 
     private ArtManager artManager;
@@ -179,317 +174,6 @@ class FlowParserTest {
             assertThatExceptionOfType(ArtParseException.class)
                     .isThrownBy(() -> parser.parse(config))
                     .withMessage("Unable to find matching parser for \"no-match\" on line 3");
-        }
-    }
-
-    @Nested
-    @DisplayName("sortAndCombineArtContexts(...)")
-    class sortAndCombine {
-
-        private List<ArtContext<?, ?, ? extends ArtObjectConfig<?>>> contexts;
-
-        @BeforeEach
-        void beforeEach() {
-            contexts = new ArrayList<>();
-        }
-
-        private ActionContext<?, ?> action() {
-            return spy(new ActionContext<>(Object.class, (o, context) -> {
-            }, new ActionConfig<>()));
-        }
-
-        private RequirementContext<?, ?> requirement() {
-            return spy(new RequirementContext<>(Object.class, (o, context) -> true, new RequirementConfig<>()));
-        }
-
-        private TriggerContext<?> trigger() {
-            return spy(new TriggerContext<>(new TriggerConfig<>()));
-        }
-
-        @Nested
-        @DisplayName("with actions as result")
-        class actions {
-
-            @Test
-            @DisplayName("should nest actions if requirements exist")
-            void shouldNestActionsIfRequirementsExist() {
-
-                ActionContext<?, ?> action = action();
-
-                contexts.addAll(Arrays.asList(
-                        requirement(),
-                        requirement(),
-                        action,
-                        action(),
-                        action()
-                ));
-
-                assertThat(parser.sortAndCombineArtContexts(contexts))
-                        .containsExactly(action)
-                        .extracting("actions.size")
-                        .contains(2);
-            }
-
-            @Test
-            @DisplayName("should create new action if new requirements exist")
-            void shouldCreateNewActionsIfNewRequirementsExist() {
-
-                ActionContext<?, ?> firstAction = action();
-                ActionContext<?, ?> secondAction = action();
-                contexts.addAll(Arrays.asList(
-                        requirement(),
-                        firstAction,
-                        requirement(),
-                        secondAction
-                ));
-
-                assertThat(parser.sortAndCombineArtContexts(contexts))
-                        .containsExactly(firstAction, secondAction)
-                        .extracting("requirements.size")
-                        .contains(1, 1);
-            }
-
-            @Test
-            @DisplayName("should discard requirements that come after the last action")
-            void shouldDiscardRequirementsAfterLastAction() {
-
-                ActionContext<?, ?> firstAction = action();
-                ActionContext<?, ?> secondAction = action();
-                contexts.addAll(Arrays.asList(
-                        requirement(),
-                        requirement(),
-                        firstAction,
-                        requirement(),
-                        secondAction,
-                        requirement(),
-                        requirement()
-                ));
-
-                assertThat(parser.sortAndCombineArtContexts(contexts))
-                        .hasSize(2)
-                        .containsExactly(firstAction, secondAction);
-            }
-
-            @Test
-            @DisplayName("should only add direct preceding requirements to action")
-            void shouldOnlyAddRelevantRequirementsToAction() {
-                ActionContext<?, ?> firstAction = action();
-                ActionContext<?, ?> secondAction = action();
-                contexts.addAll(Arrays.asList(
-                        requirement(),
-                        requirement(),
-                        requirement(),
-                        firstAction,
-                        requirement(),
-                        secondAction
-                ));
-
-                assertThat(parser.sortAndCombineArtContexts(contexts))
-                        .hasSize(2)
-                        .extracting("requirements.size")
-                        .contains(3, 1);
-            }
-
-            @Test
-            @DisplayName("should only add nested actions to preceding action")
-            void shouldAddActionsToCorrespondingAction() {
-                ActionContext<?, ?> firstAction = action();
-                ActionContext<?, ?> secondAction = action();
-                contexts.addAll(Arrays.asList(
-                        requirement(),
-                        firstAction,
-                        action(),
-                        action(),
-                        requirement(),
-                        secondAction,
-                        action()
-                ));
-
-                assertThat(parser.sortAndCombineArtContexts(contexts))
-                        .hasSize(2)
-                        .extracting("actions.size")
-                        .contains(2, 1);
-            }
-
-            @Test
-            @DisplayName("should add single action to result")
-            void shouldAddSingleAction() {
-
-                ActionContext<?, ?> action = action();
-                contexts.add(action);
-
-                assertThat(parser.sortAndCombineArtContexts(contexts))
-                        .hasSize(1)
-                        .containsExactly(action);
-            }
-        }
-
-        @Nested
-        @DisplayName("with only requirements")
-        class requirements {
-
-            @Test
-            @DisplayName("should add all requirements as flat list")
-            void shouldAddAllRequirements() {
-
-                List<RequirementContext<?, ?>> requirements = Arrays.asList(
-                        requirement(),
-                        requirement(),
-                        requirement(),
-                        requirement()
-                );
-                contexts.addAll(requirements);
-
-                assertThat(parser.sortAndCombineArtContexts(contexts))
-                        .hasSize(4)
-                        .isEqualTo(requirements);
-            }
-
-            @Test
-            @DisplayName("should add single requirement to result")
-            void shouldAddASingleRequirement() {
-
-                RequirementContext<?, ?> requirement = requirement();
-                contexts.add(requirement);
-
-                assertThat(parser.sortAndCombineArtContexts(contexts))
-                        .hasSize(1)
-                        .containsExactly(requirement);
-            }
-        }
-
-        @Nested
-        @DisplayName("with triggers")
-        class triggers {
-
-            @Test
-            @DisplayName("should register second trigger after action without requirements")
-            void shouldAddAllTriggerAsList() {
-
-                TriggerContext<?> trigger = trigger();
-                contexts.add(requirement());
-                contexts.add(requirement());
-                contexts.add(trigger);
-                contexts.add(action());
-                contexts.add(action());
-                TriggerContext<?> trigger2 = trigger();
-                contexts.add(trigger2);
-
-                assertThat(parser.sortAndCombineArtContexts(contexts))
-                        .containsExactly(trigger, trigger2)
-                        .extracting("requirements.size", "actions.size")
-                        .contains(Tuple.tuple(2, 1), Tuple.tuple(0, 0));
-            }
-
-            @Test
-            @DisplayName("should add requirements to the trigger if trigger exists after it")
-            void shouldRequirementsDirectlyToResultIfTriggerComesAfter() {
-
-                TriggerContext<?> trigger = trigger();
-                RequirementContext<?, ?> requirement = requirement();
-                contexts.add(requirement);
-                contexts.add(requirement);
-                contexts.add(trigger);
-
-                assertThat(parser.sortAndCombineArtContexts(contexts))
-                        .containsExactly(trigger)
-                        .extracting("requirements.size")
-                        .contains(2);
-            }
-
-            @Test
-            @DisplayName("should add actions directly to result if no trigger exists before it")
-            void shouldDirectlyAddActionsIfTriggerExists() {
-
-                ActionContext<?, ?> action = action();
-                TriggerContext<?> trigger = trigger();
-                contexts.add(action);
-                contexts.add(action);
-                contexts.add(trigger);
-
-                assertThat(parser.sortAndCombineArtContexts(contexts))
-                        .containsExactly(action, trigger)
-                        .extracting("actions.size")
-                        .contains(1, 0);
-
-            }
-
-            @Test
-            @DisplayName("should add actions to all matching triggers")
-            void shouldAddActionsToAllTriggers() {
-
-                TriggerContext<?> trigger1 = trigger();
-                contexts.add(trigger1);
-                contexts.add(action());
-                contexts.add(action());
-                TriggerContext<?> trigger2 = trigger();
-                contexts.add(trigger2);
-                contexts.add(action());
-
-                assertThat(parser.sortAndCombineArtContexts(contexts))
-                        .containsExactly(trigger1, trigger2)
-                        .extracting("actions.size")
-                        .contains(1, 1);
-            }
-
-            @Test
-            @DisplayName("should add requirements to all triggers that come after it")
-            void shouldAddRequirementsToAllTriggerBelow() {
-
-                contexts.add(requirement());
-                contexts.add(requirement());
-                TriggerContext<?> trigger1 = trigger();
-                contexts.add(trigger1);
-                TriggerContext<?> trigger2 = trigger();
-                contexts.add(trigger2);
-
-                assertThat(parser.sortAndCombineArtContexts(contexts))
-                        .containsExactly(trigger1, trigger2)
-                        .extracting("requirements.size")
-                        .contains(2, 2);
-            }
-
-            @Test
-            @DisplayName("two triggers in a row should get the same action")
-            void shouldGetSameActions() {
-
-                TriggerContext<?> trigger = trigger();
-                contexts.add(trigger);
-                TriggerContext<?> trigger1 = trigger();
-                contexts.add(trigger1);
-                ActionContext<?, ?> action = action();
-                contexts.add(action);
-
-                assertThat(parser.sortAndCombineArtContexts(contexts))
-                        .containsExactly(trigger, trigger1)
-                        .extracting("actions.size")
-                        .containsExactly(1, 1);
-            }
-
-            @Test
-            @DisplayName("should add requirements to actions that follow after triggers")
-            void shouldAddRequirementsToActionsFollowingTriggers() {
-
-                TriggerContext<?> trigger = trigger();
-                contexts.add(trigger);
-                contexts.add(requirement());
-                ActionContext<?, ?> action = action();
-                contexts.add(action);
-                TriggerContext<?> trigger1 = trigger();
-                contexts.add(trigger1);
-
-                assertThat(parser.sortAndCombineArtContexts(contexts))
-                        .containsExactly(trigger, trigger1);
-
-                assertThat(trigger.getActions())
-                        .containsExactly(action)
-                        .extracting(ActionContext::getRequirements)
-                        .hasSize(1);
-
-                assertThat(trigger1)
-                        .extracting("actions.size", "requirements.size")
-                        .contains(0, 0);
-            }
         }
     }
 }
